Handle edge case where backslash is used as pattern delimiter.
POSIX does not disallow backslash (\) as a pattern delimiter ed, but
it does specify that a backslash-escaped delimiter may occur in the
pattern. Therefore, just as:

```shell
ed <<'EOF'
a
hel\o
.
s;\\
EOF
```

is a valid script producing output 'helo', the following should
produce the same:

```shell
ed <<'EOF'
a
hel\o
.
s\\\
EOF
```

Prior to this commit the substitution `s\\\` was being interpreted as
`s///`, not `s/\\//p` as it should.