Handle empty insert in empty buffer.
POSIX.1-2018 states that address 0 is valid for the c command,
and "shall be interpreted as address 1 were specified".
POSIX.1-2024 removes this requirement and only "permits" address 0
for the c command.  Oddly, POSIX.1-2024 requires that address
0 be valid for the i command. A bit inconsisent.

This implementation of ed allows both address 1 and 0 where the intent
is clear. To enforce POSIX behavior, either set environment variable
POSIXLY_CORRECT or invoke ed with --traditional (-G) flag.

The following are all valid ed sessions:

    $ ed -p '*'
    *0i
    .
    *.=
    0
    *^D      <-- No modified buffer warning.
    $

    $ ed -p '*'
    *1i
    .
    *.=
    0
    *^D
    $

    $ ed -p '*'
    *0a
    .
    *.=
    0
    *^D
    $

    $ ed -p '*' <<EOF
    *1a
    .
    *.=
    0
    *^D
    $

    $ ed -vp '*'
    *a
    hello
    .
    *0c
    world
    .
    *,p
    world
    *,d
    *^D
    WARNING: Buffer modified since last write
    *^D
    $

    $ ed -vp '*'
    *a
    hello
    .
    *1c
    world
    .
    *,p
    world
    *Q
    $

In an empty buffer, the c command is not permitted, with or without an
address, since there's nothing to change.