Improve substitution extensions' POSIX compliance.
This implementation of ed includes the Berkeley repeated substitution
syntax (e.g., `s' repeats the last substitution) and extends the
available substitution modifiers. In doing so, it conflicts with the
POSIX specification:

    Any character other than <space> or <newline> can be used instead
    of a <slash> to delimit the RE and the replacement.

This is resolved by matching the substitution command with a regular
expression to disambiguate repeated substitutions from substitutions
delimited by a modifier.

For example:

    $ ed -e 's/1/_/g7p' -e 's2_202$p' -e 's2g' <(echo 1111111111111111111111)
    _111111_111111_111111_
    _111111_111111_1111110
    _111111011111101111110
    $

Here, the second expression, `s2_202$p', is a normal substitution using `2'
as delimiter while the third expression, `s2g', is a repeated substitution
starting from the second match.

There are still conflicts with POSIX.  For instance:

    $ POSIXLY_CORRECT=1 ed -e 's2g2'  <(echo abcdefg)
    abcdef
    $ ed -ve 's2g2'  <(echo abcdefg)
    /dev/fd/63: File is read-only
    ?
    script, line 1: No previous substitution
    $